requirements.txt
pandas==2.2.2
apache-airflow==2.9.3
apache-airflow-providers-google==10.15.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
